#1965 Transferred tests with BinarizeParseMojo to separate Test file
